+++ plugins/content-manager/src/main/java/com/wazuh/contentmanager/model/command/Command.java
@@ -14,7 +14,7 @@
  * You should have received a copy of the GNU Affero General Public License
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
-package com.wazuh.contentmanager.model.commandmanager;
+package com.wazuh.contentmanager.model.command;
 
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;