+++ plugins/content-manager/src/main/java/com/wazuh/contentmanager/settings/PluginSettings.java
@@ -24,7 +24,8 @@ import org.opensearch.common.settings.Setting;
 import org.opensearch.common.settings.Settings;
 import org.opensearch.core.common.settings.SecureString;
 
-import com.wazuh.contentmanager.utils.ClusterInfoHelper;
+import com.wazuh.contentmanager.utils.ClusterInfo;
+
 import reactor.util.annotation.NonNull;
 
 /** Singleton class to manage the plugin's settings. */
@@ -33,9 +34,6 @@ public class PluginSettings {
 
     /** Singleton instance. */
     private static PluginSettings INSTANCE;
-
-    /** Content Manager Plugin API path. */
-    public static final String API_BASE_URI = "/_plugins/_content_manager";
 
     /** Base Wazuh CTI URL */
     public static final String CTI_URL = "https://cti.wazuh.com";
@@ -65,6 +63,9 @@ public class PluginSettings {
     public static final Setting<SecureString> COMMAND_MANAGER_PASSWORD =
             SecureSetting.secureString("command_manager.auth.password", null);
 
+    /** Timeout of indexing operations */
+    public static final long TIMEOUT = 10L;
+
     private final String ctiBaseUrl;
     private final ClusterService clusterService;
     private final SecureString authUsername;
@@ -82,7 +83,7 @@ public class PluginSettings {
         this.authUsername = COMMAND_MANAGER_USERNAME.get(settings);
         this.authPassword = COMMAND_MANAGER_PASSWORD.get(settings);
 
-        log.debug("Settings.loaded: {}", this.toString());
+        log.debug("Settings loaded: {}", this.toString());
     }
 
     /**
@@ -129,7 +130,7 @@ public class PluginSettings {
      * @return a string with the Content Manager full URL
      */
     public String getClusterBaseUrl() {
-        return ClusterInfoHelper.getClusterBaseUrl(this.clusterService);
+        return ClusterInfo.getClusterBaseUrl(this.clusterService);
     }
 
     /**